Add files via upload
I've done all the steps except step 1, when I add the code to do this, the polygon doesn't fall or fly forward when the mouse releases it, it just stays where the mouse left it. Maybe because polygon was also created using box class so it also applies to polygon?